Engy mod fix for hives
i dont know why but if you evenr modify hive price they mess up. its
something to do with engy-mod and i guess that it will never be fixed in
normal faf (because it works there for some reason). in any case,
(hot)fixed in equilibrium but adding the line
"DifferentialUpgradeCostCalculation = false,"