novax
i add speed for 8, check it too imho its ok, but loock on it.
Also add more aoe damage, to be better against army, effekt looks biges
as only 1aoe.

and mainly that crucial mistake in code, it was ID of cybran torp
bomber, dont have clue how it was working in past, probably wasnt... :D