Seraphim MML "fix" - needs attention when moving to stable!
i swear if anyone makes a single complaint about this then i will go to
their house and force them to relive what i did to make this work.
in other news Ithilis no longer owns a dog.

this WILL break when being released - you need to change the projectile
directory in the script or it wont find it when put on stable.