some video + mex e price cost
it contain soem ncie video production, to changelog, is about half that
i have in plane,   and also small change on mex energy cost. mainly for
drain reason,   in past was calculate it also with drain for produce
mass, to drain nicer number while upgrade, but its not wort it to see
ugly number on building list.