added better jamming script for uef jammer.
it now works completely as intended, jamming off by default and draining
25 energy. stealth drains 125 and they add up correctly. spinner stops
spinning only when both are off, and ect.

the only thing i couldnt get to work is the strategic icon shows as
paused/disabled when either thing is turned off, but i want it to show
only when both are turned off, dunno how to do that.