fix mobile shields coming too close to enemy
this was some sort of bug in the past, and now its been fixed. the
shields should stop at the best distance from the enemy so they are just
behind your units and out of range of the dangerous stuff.
as a slight bonus there are tiny visual improvements to the aeon hover
shield - it now hovers upright and wobbles a bit more :3