RNB - Clean Torp launcher stealth
bscly removing all the useless bits. the maxspeed is needed and so is
the motion type - it cant be none - then pick sth that fits - for harms
its amphibious so it stays on the seabed.

this can be used to make any building sneaky. probably a good idea on
stealth generators?