sera acu regen aura
fixed bug that cause wierd behaviour when is near advanced regen aura.

check how they fix it in vanila, and this is only differences what i
found and make perfecty sance that this is responsible for it.  (dont
test it yet)

theoreticly this should be somewhere in lua file too, but nearly sure
that eq dont touch this parth .