changelog and mod info optimalization
this loock chaotic, becsaue it is!
Some crazy noob click revert on torrent and then change opinion, but it
was in comit, and then by mistake click to add all changes so it fuck up
torrent. So previous revert and this comit is basicly to repair previous
shitnes what was done by some crazy noob....