Fix sera acu adv gun removal bug
aha more crazy ithilis code exploits!

when something is defined as a local it means it can only be found
inside that bit its defined in. this means you need to redefine what
bpDisrupt is. also try reading the error log instead of saying "its some
kind of bug" xDD