Fix Aeon TMD not hitting tmls flying way over it
ok so this creates additional collision detectors above and below it, so
its 3 spheres instead of just one, which means it can catch missiles a
lot higher than it.

this means we can mess with the projectile a lot more now to look cool
as well. shooting up then hovering at a set height?