Improve sacrifice by refunding wasted mass in wrecks
the leftover mass due to differences in the ratios of the units is now
dumped on the ground in the form of a wreck. it doesnt create mass out
of thin air and doesnt give negative mass values. in the case of energy
being wasted, that does not get refunded.