Fix Ythotha AA cannons missing
they were pretty retarded but it should be ok now - added LeadTarget =
true, and also UseFiringSolutionInsteadOfAimBone has to be used because
its model is fucked, needs to be fixed by a modeller or use this. Also
no idea why muzzle salvo size was 3 on there. also for this reason we
had some absurd 250 dps on the unit db when its actually more like 80