Fix Engies not being in vet table and add logging
so if a unit isnt in the vet table it doesnt track vet. this is
obviously bad if it kills something.
tbh i think that this very idea that not every unit tracks vet is
totally absurd, but w.e.

anyway now engies and sparkies are added as raider types, so they need
minimum vet or w.e.