Unhook AdjacencyBuffs so it doesnt fill duplicate values
ok so it worked and all, but this way it replaces the original table
instead of trying to fill it again after its done. this way the
duplicate value detected warning doesnt show up, which is cool

its no longer hooked which is kinda bad but really its only ever going
to contain numbers and that table filler which isnt exactly something
you can hook anyway.