Update UEF camera code + fix tmls
so there was an issue that the code in there was causing proj to return
as nil value inside the rest of the weapon scripts which is bad if they
want to do something to it.

dont trust gpg i spose.

this caused the ammo to not be subtracted for counted projectiles, so
you could get tmls for free.

now thats fixed, and also a lot less file changes are required so thats
cool. however when adding new weapons or changing their order, you will
break the new code since it relies on a number index in the weapon code.
i dont see that happening often though so thats ok.