Rhino sound
http://recordit.co/RsnsV2mnxB

Rhino do damage every 0,1s instead of 0,2. But sound is every 1s. Now
work coreclty and not hack system with low rof. Check gif that show how
it work, and that any trash that you was talking about is not true. Easy
as that. Problem was that on beam it do dmg as when its single weaopn
and divide on 2 as it was before.