Make ACU Drones upgrade with ACU tech level
so now if you get t2 with acu your drones get 10 br each, and 20 if its
t3. you dont keep the upgraded drones if you remove the tech upgs.

there is also a cool effect on the drones whenever they are created or
their buildpower changes.

but most importantly, we now have cd1 on the first upgrade, and cd2 on
the second drone! now the buddies can be together!

(cd2 is for scus but this is way cooler isnt it)