Improve Scathis Usablity
the scathis still needs some improvement, so it got a faster pack/unpack
animation, more hp, aoe and a better fire cycle but with the same dps.

the effectiveness vs armies should be increased and it should be safer.
with proper application it can be used vs shields in an offensive manner
too.