Fix SACU upgrade vet bug
the issue is if you upgrade a sacu or downgrade a preset they still give
the same vet. thats bad and we fix it by saving its price on create then
changing it dynamically.

this also leads to other interesting things that could be done with
dynamic repair costs, ect.

however, wrecks are unaffected, thats for later.

also there is a minor performance improvement in OnCreate which is why
we hook the whole thing.