lab buff
increse range of lab to 15 (13)
We was discuse that t1 tanks should have more dps instead of hp. And
labs are already on this position,  but are missing range. It is same
problem with t1 tanks but that is too big change.

And while it should be realy cancer to hit lab with higer range its also
biger hitboxes to make less problem with dodge.