Reduce gpg-like behaviour for planes refueling
if you toggle or spawn a lot at a time this can happen.
this random function spreads out the time the planes try to refuel so
they go for the same platform less often,
but this function still works best when you never toggle it and build
planes without cheating them in.