Rework sera sniper bot
now no speed penalty, no more cloak, more hp and more range on the
sniper mode. pretty good when microed well, can be stopped by titans and
stuff, at an advantage when enemy is afk and weak if you dont run 100%
of the time but stop to shoot, which is fine.