Reduce mass storage cost
i hate to admit this, but it seems i fucked up when making this. while i
can understand why i did this, it makes an awful lot more sense to do it
like this xD

this way the storages dont slow down your ecoing, and are very buffed
indeed for storage purposes so you can get one early since theyre so
cheap.

and i also hate to admit this but thanks to bh for bringing my attention
to this.