Fix Notha air behaviour
Thanks to Evildrew for pointing this one out.
The Notha had NeedPrep enabled on all its weapons which led to the
frequent calling of OnGotTarget and thats not needed behaviour since it
only applies to the bomb weapon, and that works on engine code anyway.
For some reason it didnt always call OnLostTarget like this, which led
to the multipliers to be stuck in bombing mode and that caused the
issues.