Optimise SeabedReveal
this was cycling through two tables, now it cycles through only one.
also in case a unit has a nil value on its watervision and has an
expired timestamp, we reset it to the blueprint value or our best guess
for it, so they have no chance to clog up the table. not that it should
happen in any case.