Overhaul Ingame stats view
Technically, this was a fix. But so much of it was wrong it might as
well be a rewrite.
Fixed things:
ProjectilesPerOnFire no longer used - its not a real stat
Reload time calculated to the nearest tick and not to the nearest
whatever
Charge time and rackreload time used correctly
Fixed duplicate weapon detection
Nukes give real stats now and not 0 damage
Weapons firing salvoes now give correct damage/salvo
Fixed not calculating damage over time
Fixed not calculating initial damage in DOT
Fixed not calculating beam damage duration