Fix Torpedo bomber dropping strangeness
spaghetti discovered fin the aeon and sera bombers.

i dunno what the purpose was of all that wierd code but it led to
overshooting because the projectiles were not tuned correctly, or it
would have led to the original issue of hitting the shore when
attacking. so instead we make them work like the uef/cyb bombers and
call it a day. more consistent like that anyway.