Protect EQ from shield balance changes
So, it was always 15% for everything due to some bug. Things were
balanced to that 15%. Since there is no obvious need to buff/nerf things
in shields with eq, it will be staying like that.

Makes the understanding simpler as well, can always buff/nerf shields in
some other way.