Nerf UEF lab
well this was kinda planned for a few months, just waiting for some
backup games to show it, but we can always say that officially it was
due to BHs excellent constructive feedback on the matter.

also ithilis misleading with wrong calculations is still haunting us to
this day!

now uef labs are still the best just not as good.