Redo autorefuel code for performance + consistency
now its done a bit like the watervision reset thread, with a global air
manager that ensures all the planes are sent to the right platforms
correctly. this avoids creating a bunch of threads and should be less
performance hungry.

(however a bunch of threads is technically more robust so theres that,
too)