Rebalance Cybran Cruiser
it now has slightly more convenient projectiles, which makes it tougher
against torpedo bombers and low tech stuff, but a little lower DPS now
so its about as good as the aeon cruiser against t3 things, and pretty
strat proof now which is good. overall its not the worst cruiser for AA
anymore for sure.