Improve UEF T3 MAA targetting
now it can hit things like spyplanes better. a lot better.

its kinda weird how the projectiles work actually. having both tracking
and leading target is actually a shit idea, but combined with a low turn
rate the tracking fucks it up just enough to not ruin it in some
situations, but when the aa shoots behind its target due to bad engine
code it fixes the lead. funny how these things work xD

the real magic here is in the DetonatesAtTargetHeight which means that
normally the aa misses just enough to not hit but this makes it into a
true monster with one line of code :>