Buff Aeon frigate torp defense
now the defense is better than before, it wasnt really that bad
mechanically but had overkill issues due to low projectile speed. this
is alleviated with a higher ROF as opposed to increasing the speed to
keep the changes localized. perhaps it is better to increase the speed,
but that needs further testing. also, the lifetime is increased, and the
frigate now alternates between its two muzzles as opposed to just one
every time.