Remove ability to call functions with target prios
the assumption i made previously was that it wasnt possible to run code
that is inside pcall().
apparently thats incorrect, and to adjust for that we check for a bit of
structure first before compiling the string. this means that anything
that contains a function or method fails the check since brackets are
required for that to work.

im sure glad i made this code modular.

also fixed a horrid typo in the function name xD