Sera cruiser submersibility
Add submersibility to cruiser, to fit with destroyer. Units be design
looks good for it, and it add only hide/escape toll while it not fire in
sub form.

It increase diversity. Think about also add it into frigate, but this
can create future problems to be too strong, while not leave reclaim on
early fights and there will be none to repair it. So i only keep this
note.