Add stealth to transported units
The burning question in me is this:
What, precisely, is the proportion of people who are in favour, or
against this idea?

After all, wouldnt it be a shame if some cancer like this makes it in
and no one even wanted it? Wouldnt it be just fantastic to i dunno,
check if anyone likes this first? Ah well, not my problem now.

To be clear, this exists because Ithilis made me do it.